python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "authy"
version = "0.1.0"
description = "Multi-provider authentication framework supporting local, Google, M365, and SSO"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["authentication", "oauth", "oidc", "saml", "google", "m365", "sso"]
dependencies = [
    "authlib>=1.3.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "httpx>=0.27.0",
    "python3-saml>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
